chore: Moves BootstrapSliderWrapper to own folder (#13460)
