+++ superset-frontend/src/components/BootstrapSliderWrapper/BootstrapSliderWrapper.less
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-@import '../../stylesheets/less/variables.less';
+@import '../../../stylesheets/less/variables.less';
 
 .BootstrapSliderWrapper .slider-selection {
   background: @gray-bg;
